requirements.txt
requests==2.26.0
pytest==6.2.5
vcrpy==4.1.1
